+++ frontend/src/modules/2DViewer/DataProviderFramework/visualization/makeStatisticalSurfaceLayer.ts
@@ -11,7 +11,7 @@ import {
     type StatisticalSurfaceData,
     type StatisticalSurfaceSettings,
     SurfaceDataFormat,
-} from "../customDataProviderImplementations/StatisticalSurface";
+} from "../customDataProviderImplementations/StatisticalSurfaceProvider";
 
 function calcBoundsForRotationAroundUpperLeftCorner(surfDef: SurfaceDef_api): [number, number, number, number] {
     const width = (surfDef.npoints_x - 1) * surfDef.inc_x;
