+++ frontend/src/modules/_shared/DataProviderFramework/dataProviders/registerAllDataProviders.ts
@@ -1,7 +1,7 @@
 import { DataProviderRegistry } from "./DataProviderRegistry";
 import { DataProviderType } from "./dataProviderTypes";
-import { DrilledWellTrajectories } from "./implementations/DrilledWellTrajectories";
-import { DrilledWellborePicks } from "./implementations/DrilledWellborePicks";
+import { DrilledWellTrajectoriesProvider } from "./implementations/DrilledWellTrajectoriesProvider";
+import { DrilledWellborePicksProvider } from "./implementations/DrilledWellborePicksProvider";
 
-DataProviderRegistry.registerDataProvider(DataProviderType.DRILLED_WELLBORE_PICKS, DrilledWellborePicks);
-DataProviderRegistry.registerDataProvider(DataProviderType.DRILLED_WELL_TRAJECTORIES, DrilledWellTrajectories);
+DataProviderRegistry.registerDataProvider(DataProviderType.DRILLED_WELLBORE_PICKS, DrilledWellborePicksProvider);
+DataProviderRegistry.registerDataProvider(DataProviderType.DRILLED_WELL_TRAJECTORIES, DrilledWellTrajectoriesProvider);
